split tests, write object property parsers
